backend/requirements.txt
flask>=2.3.3
flask-cors==3.0.10
google-generativeai==0.3.0
python-dotenv==0.19.0
yt-dlp==2023.12.30
openai-whisper==20231117
ffmpeg-python==0.2.0
werkzeug>=3.0.0 